chore: Move files and functions around
