+++ src/casing.test.ts
@@ -1,58 +1,7 @@
-import type * as Subject from './casing'
 import * as subject from './casing'
 
-const weirdString = ' someWeird-cased$*String1986Foo Bar W_FOR_WUMBO' as const
-type WeirdString = typeof weirdString | 'dont.distribute unions'
-
-namespace TypeTransforms {
-  type test = Expect<
-    Equal<
-      Subject.DelimiterCase<WeirdString, '%'>,
-      | 'some%Weird%cased%$*%String%1986%Foo%Bar%W%FOR%WUMBO'
-      | 'dont%distribute%unions'
-    >
-  >
-  type test1 = Expect<
-    Equal<
-      Subject.CamelCase<WeirdString>,
-      'someWeirdCased$*String1986FooBarWForWumbo' | 'dontDistributeUnions'
-    >
-  >
-  type test2 = Expect<
-    Equal<
-      Subject.PascalCase<WeirdString>,
-      'SomeWeirdCased$*String1986FooBarWForWumbo' | 'DontDistributeUnions'
-    >
-  >
-  type test3 = Expect<
-    Equal<
-      Subject.KebabCase<WeirdString>,
-      | 'some-weird-cased-$*-string-1986-foo-bar-w-for-wumbo'
-      | 'dont-distribute-unions'
-    >
-  >
-  type test4 = Expect<
-    Equal<
-      Subject.SnakeCase<WeirdString>,
-      | 'some_weird_cased_$*_string_1986_foo_bar_w_for_wumbo'
-      | 'dont_distribute_unions'
-    >
-  >
-  type test5 = Expect<
-    Equal<
-      Subject.ConstantCase<WeirdString>,
-      | 'SOME_WEIRD_CASED_$*_STRING_1986_FOO_BAR_W_FOR_WUMBO'
-      | 'DONT_DISTRIBUTE_UNIONS'
-    >
-  >
-  type test6 = Expect<
-    Equal<
-      Subject.TitleCase<WeirdString>,
-      | 'Some Weird Cased $* String 1986 Foo Bar W For Wumbo'
-      | 'Dont Distribute Unions'
-    >
-  >
-}
+export const weirdString =
+  ' someWeird-cased$*String1986Foo [Bar] W_FOR_WUMBO...' as const
 
 describe('capitalize', () => {
   test('it does nothing with a string that has no char at the beginning', () => {
@@ -88,14 +37,16 @@ describe('uncapitalize', () => {
 
 describe('casing functions', () => {
   test('toUpperCase', () => {
-    const expected = ' SOMEWEIRD-CASED$*STRING1986FOO BAR W_FOR_WUMBO' as const
+    const expected =
+      ' SOMEWEIRD-CASED$*STRING1986FOO [BAR] W_FOR_WUMBO...' as const
     const result = subject.toUpperCase(weirdString)
     expect(result).toEqual(expected)
     type test = Expect<Equal<typeof result, typeof expected>>
   })
 
   test('toLowerCase', () => {
-    const expected = ' someweird-cased$*string1986foo bar w_for_wumbo' as const
+    const expected =
+      ' someweird-cased$*string1986foo [bar] w_for_wumbo...' as const
     const result = subject.toLowerCase(weirdString)
     expect(result).toEqual(expected)
     type test = Expect<Equal<typeof result, typeof expected>>
